Added user search API.

+++ src/main/java/com/example/demo/backend/UserBusinessLogic.java
@@ -1,7 +1,7 @@
 package com.example.demo.backend;
 
 
-import static com.example.demo.backend.errorcodes.SqlErrorCodes.SQL_CODE_DUPLICATE_KEY_ERROR;
+import static com.example.demo.backend.errorcodes.SqlErrorCodes.UNIQUE_VIOLATION;
 
 
 
@@ -14,28 +14,42 @@ import com.example.demo.common.exceptions.UserBusinessLogicException;
 import com.example.demo.data.access.JdbcUserDao;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
 
-@Component
+@Service
 public class UserBusinessLogic {
-
+  private ResponseUsers ures;
 
   @Autowired
   private JdbcUserDao udao;
-  private ResponseUsers ures;
+
+
+  public ResponseUsers getAllUsers() throws DaoException, UserBusinessLogicException {
+    ures =  ResponseUsers.builder().build();
+    List<User> users = udao.getAllUsers();
+    if (users.isEmpty()) {
+      throw new UserBusinessLogicException(Messages.USER_NOT_EXISTING);
+    } else {
+      ures.getMessageHeader().setMessage(Messages.USER_FOUND);
+    }
+    ures.setUsers(users);
+    return ures;
+  }
 
 
   public ResponseUsers addUser(User user) throws DaoException, UserBusinessLogicException {
     ures = ResponseUsers.builder().build();
     try {
-      List<User> users = udao.insertBookUser(user);
-      ures.setUsers(users);
-      ures.getMessageHeader().setMessage(Messages.USER_INSERTED);
+      int updated = udao.insertBookUser(user);
+      if (updated == 0) {
+        throw new UserBusinessLogicException(Messages.UPDATE_FAILED_USER);
+      } else {
+        ures.getMessageHeader().setMessage(Messages.USER_INSERTED);
+      }
       return ures;
     } catch (DaoException e) {
-      if (e.getSqlCode().equals(SQL_CODE_DUPLICATE_KEY_ERROR)) {
+      if (e.getSqlCode().equals(UNIQUE_VIOLATION)) {
         throw new UserBusinessLogicException(Messages.USER_DUPLICATE);
       }
       throw e;
@@ -45,14 +59,19 @@ public class UserBusinessLogic {
 
   public ResponseUsers removeUser(Integer userId) throws DaoException, UserBusinessLogicException {
     ures = ResponseUsers.builder().build();
-    //Even better, must check whether the user is not borrowing any book.
 
-    int update = udao.deleteBookUser(userId);
-    if (update == 0) {
-      throw new UserBusinessLogicException(Messages.USER_NOT_EXISTING);
+    List<String> booksBorrowed = udao.getBorrowedBookTitles(userId);
+    if(booksBorrowed.isEmpty()) {
+      int update = udao.deleteBookUser(userId);
+      if (update == 0) {
+        throw new UserBusinessLogicException(Messages.USER_NOT_EXISTING);
+      } else {
+        ures.getMessageHeader().setMessage(Messages.USER_DELETED);
+      }
     } else {
-      ures.getMessageHeader().setMessage(Messages.USER_DELETED);
+      throw new UserBusinessLogicException(Messages.USER_CANNOT_BE_DELETED);
     }
+
     return ures;
   }
